refactor: - move constant from type args in parsing vector - move tests to separate mod
