+++ src/lib.rs
@@ -1,8 +1,8 @@
 //! NFS Mamont - A Network File System (NFS) server implementation in Rust.
 
 mod mount;
-mod nfsv3;
-mod parser;
+pub mod nfsv3;
+pub mod parser;
 mod read_task;
 mod rpc;
 pub mod vfs;
